formatting

+++ timeevents/file_clean.py
@@ -9,8 +9,7 @@ def delete_old(path: str):
   for file in os.scandir(path):
     file_path = os.path.join(path, file.name)
 
-    if file.is_dir():
-      delete_old(file_path)
+    if file.is_dir(): delete_old(file_path)
     elif date.fromtimestamp(os.path.getmtime(file_path)) < time:
       print(f'deleting {file_path}')
       os.unlink(file_path)
@@ -22,7 +21,7 @@ START_NOW = True
 def on_tick(self):
   now = date.today().strftime('%m-%d')
 
-  if self.settings['lastFileClear'] == now:
+  if self.settings.lastFileClear == now:
     print('Already ran file deletion today')
     return
